+++ cmd/terramate/e2etests/cloud/run_cloud_drift_test.go
@@ -248,6 +248,30 @@ func TestCLIRunWithCloudSyncDriftStatus(t *testing.T) {
 			},
 		},
 		{
+			name:   "basic drift sync with uppercase stack id",
+			layout: []string{"s:stack:id=STACK"},
+			cmd: []string{
+				HelperPath, "exit", "2",
+			},
+			want: want{
+				drifts: expectedDriftStackPayloadRequests{
+					{
+						DriftStackPayloadRequest: cloud.DriftStackPayloadRequest{
+							Stack: cloud.Stack{
+								Repository:    normalizedTestRemoteRepo,
+								DefaultBranch: "main",
+								Path:          "/stack",
+								MetaName:      "stack",
+								MetaID:        "stack",
+							},
+							Status:   drift.Drifted,
+							Metadata: expectedMetadata,
+						},
+					},
+				},
+			},
+		},
+		{
 			name: "only stacks inside working dir are synced",
 			layout: []string{
 				"s:parent:id=parent",